add search in header input
